commander: Change commander packet to be a float instead. Rework ordering of when thrust command is sent.

+++ src/modules/interface/commander.h
@@ -47,7 +47,7 @@ struct CommanderCrtpValues
   float roll;       // deg
   float pitch;      // deg
   float yaw;        // deg
-  uint16_t thrust;
+  float thrust;
 } __attribute__((packed));
 
 void commanderInit(void);
